requirements.txt
cel-python>=0.1.5
lark>=0.12.0
selenium>=4.1.0
tabulate>=0.8.9
z3-solver>=4.8.11.0